fix: 更新

+++ frontend/src/contracts/abis/manager.ts
@@ -17,7 +17,7 @@ export const managerAbi = [
     inputs: [
       { internalType: 'address', name: 'poolAddress_', type: 'address' },
       { internalType: 'bool', name: 'zeroForOne', type: 'bool' },
-      { internalType: 'int256', name: 'amountSpecified', type: 'int256' },
+      { internalType: 'uint256', name: 'amountSpecified', type: 'uint256' },
       { internalType: 'bytes', name: 'data', type: 'bytes' },
     ],
     name: 'swap',
@@ -39,7 +39,7 @@ export const managerAbi = [
   {
     inputs: [
       { internalType: 'int256', name: 'amount0Delta', type: 'int256' },
-      { internalType: 'int256', name: 'amount1Delta', type: 'int24' },
+      { internalType: 'int256', name: 'amount1Delta', type: 'int256' },
       { internalType: 'bytes', name: 'data', type: 'bytes' },
     ],
     name: 'uniswapV3SwapCallback',
